+++ src/components/level/toolbars/EditToolbar/selection-editors/gradient.tsx
@@ -1,0 +1,14 @@
+import { FC } from "react";
+import { svgs } from "ui/icon";
+import { SelectionEditor, SelectionEditorProps } from "./_types";
+
+const GradientEditor: FC<SelectionEditorProps> = ({ onSubmit }) => (
+  <div>...</div>
+);
+
+export const gradient: SelectionEditor = {
+  title: "Gradient",
+  // TODO: icon
+  icon: <svgs.Pencil />,
+  Component: GradientEditor,
+};