+++ src/components/level/toolbars/EditToolbar/selection-editors/maze.tsx
@@ -1,0 +1,14 @@
+import { FC } from "react";
+import { svgs } from "ui/icon";
+import { SelectionEditor, SelectionEditorProps } from "./_types";
+
+const MazeEditor: FC<SelectionEditorProps> = ({ onSubmit }) => <div>...</div>;
+
+export const maze: SelectionEditor = {
+  title: "Maze",
+  // TODO: icon
+  icon: <svgs.Pencil />,
+  cannotWorkWhy: (s) =>
+    s.width < 5 || s.height < 5 ? <>at least 5x5</> : null,
+  Component: MazeEditor,
+};