+++ src/components/level/toolbars/EditToolbar/selection-editors/replace.tsx
@@ -1,0 +1,14 @@
+import { FC } from "react";
+import { svgs } from "ui/icon";
+import { SelectionEditor, SelectionEditorProps } from "./_types";
+
+const ReplaceEditor: FC<SelectionEditorProps> = ({ onSubmit }) => (
+  <div>...</div>
+);
+
+export const replace: SelectionEditor = {
+  title: "Replace",
+  // TODO: icon
+  icon: <svgs.Pencil />,
+  Component: ReplaceEditor,
+};