+++ src/components/level/toolbars/EditToolbar/selection-editors/rnd.tsx
@@ -1,0 +1,12 @@
+import { FC } from "react";
+import { svgs } from "ui/icon";
+import { SelectionEditor, SelectionEditorProps } from "./_types";
+
+const RndEditor: FC<SelectionEditorProps> = ({ onSubmit }) => <div>...</div>;
+
+export const rnd: SelectionEditor = {
+  title: "Random",
+  // TODO: icon
+  icon: <svgs.Pencil />,
+  Component: RndEditor,
+};