+++ src/drivers/supaplex/driver.ts
@@ -1,7 +1,7 @@
 import { DAT, MPX, SP } from "./formats";
 import { LevelConfigurator } from "./LevelConfigurator";
 import { applyLocalOptions, LevelLocalOptions } from "./LevelLocalOptions";
-import { ISupaplexDriver } from "./types";
+import { ISupaplexDriver, ISupaplexLevelRegion } from "./types";
 import { Tile } from "./Tile";
 import { tiles } from "./tiles";
 
@@ -22,4 +22,6 @@ export const SupaplexDriver: ISupaplexDriver = {
     sp: SP,
   },
   defaultFormat: "dat",
+  tempLevelFromRegion: (r) =>
+    MPX.createLevel(r.tiles).pasteRegion(0, 0, r as ISupaplexLevelRegion),
 };