+++ src/ui/button/Toolbar/Toolbar.tsx
@@ -5,14 +5,20 @@ import cl from "./Toolbar.module.scss";
 
 interface Props extends ContainerProps {
   withBG?: boolean;
+  isMenu?: boolean;
 }
 
 export const Toolbar = forwardRef<HTMLDivElement, PropsWithChildren<Props>>(
-  ({ withBG = true, className, children, ...rest }, ref) => (
+  ({ withBG = true, isMenu = false, className, children, ...rest }, ref) => (
     <div
       {...rest}
       ref={ref}
-      className={cn(cl.root, withBG && cl._withBG, className)}
+      className={cn(
+        cl.root,
+        withBG && cl._withBG,
+        isMenu && cl._menu,
+        className,
+      )}
     >
       {children}
     </div>