+++ CHANGELOG.md
@@ -6,6 +6,9 @@
   more than 256 infotrons.
 - Add: Pen tool with shapes 1x2 and 2x1 with Double Chips support. This will be
   enhanced in further versions.
+- Add: Selection edit tools. Select a region of level body, and the new dropdown
+  button will allow to apply to the selection the following tools:
+  - Flip.
 - Enh: Toolbars layout now is more adaptive to be more compact together.
 - Enh: Special Ports buttons in tiles toolbar was merged with corresponding
   Regular Ports. So now there are 4 fewer buttons.